feat: implement API handlers for clubs and members with CORS and logging middleware
