+++ backend/go.mod
@@ -7,6 +7,7 @@ require github.com/google/uuid v1.6.0
 require github.com/lib/pq v1.10.9
 
 require (
+	github.com/gorilla/mux v1.8.1 // indirect
 	github.com/jackc/pgpassfile v1.0.0 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
 	github.com/jackc/pgx/v5 v5.7.2 // indirect
