+++ backend/handlers/middlewares.go
@@ -1,0 +1,56 @@
+package handlers
+
+import (
+	"log"
+	"net/http"
+	"time"
+)
+
+type responseWriter struct {
+	http.ResponseWriter
+	status int
+}
+
+func (rw *responseWriter) WriteHeader(code int) {
+	rw.status = code
+	rw.ResponseWriter.WriteHeader(code)
+}
+
+func CorsMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Access-Control-Allow-Origin", "*")
+		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
+		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
+
+		// Handle preflight request
+		if r.Method == "OPTIONS" {
+			w.WriteHeader(http.StatusOK)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	})
+}
+
+func LoggingMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		start := time.Now()
+
+		// Create custom response writer to capture status code
+		rw := &responseWriter{
+			ResponseWriter: w,
+			status:         http.StatusOK, // Default status
+		}
+
+		next.ServeHTTP(rw, r)
+
+		duration := time.Since(start)
+		log.Printf(
+			"endpoint=%s method=%s status=%d duration=%v",
+			r.URL.Path,
+			r.Method,
+			rw.status,
+			duration,
+		)
+	})
+}
